[renderpack] Rename most usages of the word 'shaderpack' to renderpack
