+++ include/nova_renderer/constants.hpp
@@ -18,7 +18,7 @@ namespace nova::renderer {
 
     constexpr mem::Bytes PER_FRAME_MEMORY_SIZE = 2_mb;
 
-    constexpr const char* RENDERPACK_DIRECTORY = "shaderpacks";
+    constexpr const char* RENDERPACK_DIRECTORY = "renderpacks";
     constexpr const char* MATERIALS_DIRECTORY = "materials";
     constexpr const char* SHADERS_DIRECTORY = "shaders";
     constexpr const char* RENDERPACK_DESCRIPTOR_FILE = "renderpack.json";
