+++ include/nova_renderer/nova_renderer.hpp
@@ -82,17 +82,17 @@ namespace nova::renderer {
         ~NovaRenderer();
 
         /*!
-         * \brief Loads the shaderpack with the given name
+         * \brief Loads the renderpack with the given name
          *
-         * This method will first try to load from the `shaderpacks/` folder (mimicking Optifine shaders). If the
-         * shaderpack isn't found there, it'll try to load it from the `resourcepacks/` directory (mimicking Bedrock
-         * shaders). If the shader can't be found at either place, a `nova::resource_not_found` exception will be thrown
+         * This method will first try to load from the `renderpacks/` folder. If Nova can't find the renderpack there, it will try to load
+         * it from the `shaderpacks/` directory (mimicking Optifine shaders). If the renderpack isn't found there, it'll try to load it from
+         * the `resourcepacks/` directory (mimicking Bedrock shaders)
          *
-         * Loading a shaderpack will cause a stall in the GPU. Nova will have to wait for all in-flight frames to finish, then replace the
-         * current shaderpack with the new one, then start rendering. Replacing the shaderpack might also require reloading all chunks, if
-         * the new shaderpack has different geometry filters then the current one
+         * Loading a renderpack will cause a stall in the GPU. Nova will have to wait for all in-flight frames to finish, then replace the
+         * current renderpack with the new one, then start rendering. Replacing the renderpack might also require reloading all chunks, if
+         * the new renderpack has different geometry filters then the current one
          *
-         * \param renderpack_name The name of the shaderpack to load
+         * \param renderpack_name The name of the renderpack to load
          */
         void load_renderpack(const rx::string& renderpack_name);
 
@@ -218,7 +218,7 @@ namespace nova::renderer {
         rx::memory::allocator* global_allocator;
 
         /*!
-         * \brief Holds all the object loaded by the current rendergraph
+         * \brief Holds all the object loaded by the current renderpacks
          */
         rx::memory::allocator* renderpack_allocator;
 
@@ -276,9 +276,9 @@ namespace nova::renderer {
             PipelineMetadata metadata;
         };
 
-        bool shaderpack_loaded = false;
+        bool renderpacks_loaded = false;
 
-        rx::concurrency::mutex shaderpack_loading_mutex;
+        rx::concurrency::mutex renderpacks_loading_mutex;
 
         rx::optional<renderpack::RenderpackData> loaded_renderpack;
