+++ src/rhi/vulkan/vulkan_utils.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "nova_renderer/rhi/command_list.hpp"
-#include "nova_renderer/shaderpack_data.hpp"
+#include "nova_renderer/renderpack_data.hpp"
 
 // idk maybe this header is included in places that already include Vulkan? Either way I want this include here and not anywhere else
 // ReSharper disable once CppUnusedIncludeDirective
