--- tests/resources/shaderpacks/DefaultShaderpack/materials/final.mat
@@ -1,16 +1,0 @@
-{
-    "name": "final",
-    "filter": "geometry_type::fullscreen_quad",
-    "passes": [
-        {
-            "name": "main",
-            "pipeline": "Final",
-            "bindings": [
-                {
-                    "variable": "per_model_uniforms",
-                    "resource": "NovaModelMatrixBuffer"
-                }
-            ]
-        }
-    ]
-}
