--- tests/resources/shaderpacks/DefaultShaderpack/materials/gbuffers_terrain.mat
@@ -1,16 +1,0 @@
-{
-    "name": "gbuffers_terrain",
-    "filter": "geometry_type::block AND not_transparent",
-    "passes": [
-        {
-            "name": "main",
-            "pipeline": "gbuffers_terrain",
-            "bindings": [
-                {
-                    "variable": "per_model_uniforms",
-                    "resource": "NovaModelMatrixBuffer"
-                }
-            ]
-        }
-    ]
-}
