--- tests/resources/shaderpacks/DefaultShaderpack/materials/gui.mat
@@ -1,16 +1,0 @@
-{
-    "name": "gui",
-    "filter": "geometry_type::gui",
-    "passes": [
-        {
-            "name": "main",
-            "pipeline": "gui",
-            "bindings": [
-                {
-                    "variable": "per_model_uniforms",
-                    "resource": "NovaModelMatrixBuffer"
-                }
-            ]
-        }
-    ]
-}
