--- tests/resources/shaderpacks/DefaultShaderpack/materials/gui.pipeline
@@ -1,28 +1,0 @@
-{
-    "name": "gui",
-    "pass": "Forward",
-    "depthFunc": "LessEqual",
-    "states": [
-        "DisableAlphaWrite"
-    ],
-    "vertexShader": "shaders/gui.vert",
-    "fragmentShader": "shaders/gui.frag",
-    "vertexFields": [
-        {
-            "name": "position_in",
-            "field": "Position"
-        },
-        {
-            "name": "uv_in",
-            "field": "UV0"
-        },
-        {
-            "name": "color_in",
-            "field": "Color"
-        },
-        {
-            "name": "virtual_texture_id_in",
-            "field": "VirtualTextureId"
-        }
-    ]
-}
