--- tests/resources/shaderpacks/DefaultShaderpack/materials/gui_background.mat
@@ -1,16 +1,0 @@
-{
-    "name": "gui_background",
-    "filter": "geometry_type::gui_background",
-    "passes": [
-        {
-            "name": "main",
-            "pipeline": "gui",
-            "bindings": [
-                {
-                    "variable": "per_model_uniforms",
-                    "resource": "NovaModelMatrixBuffer"
-                }
-            ]
-        }
-    ]
-}
