--- tests/resources/shaderpacks/DefaultShaderpack/materials/gui_text.mat
@@ -1,16 +1,0 @@
-{
-    "name": "gui_text",
-    "filter": "geometry_type::text",
-    "passes": [
-        {
-            "name": "main",
-            "pipeline": "gui",
-            "bindings": [
-                {
-                    "variable": "per_model_uniforms",
-                    "resource": "NovaModelMatrixBuffer"
-                }
-            ]
-        }
-    ]
-}
