--- tests/resources/shaderpacks/DefaultShaderpack/rendergraph.json
@@ -1,36 +1,0 @@
-{
-    "passes": [
-        {
-            "name": "Forward",
-            "textureOutputs": [
-                {
-                    "name": "LitWorld",
-                    "clear": true
-                }
-            ],
-            "depthTexture": {
-                "name": "DepthBuffer",
-                "clear": true
-            },
-            "bufferInputs": [
-                "NovaMegaMesh_Vertices",
-                "NovaMegaMesh_Indices"
-            ]
-        },
-        {
-            "name": "Final",
-            "textureInputs": [
-                "LitWorld"
-            ],
-            "textureOutputs": [
-                {
-                    "name": "NovaBackbuffer",
-                    "clear": false
-                }
-            ]
-        }
-    ],
-    "builtin_passes": [
-        "NovaUI"
-    ]
-}
