--- tests/resources/shaderpacks/DefaultShaderpack/resources.json
@@ -1,29 +1,0 @@
-{
-    "textures": [
-        {
-            "name": "LitWorld",
-            "format": {
-                "pixelFormat": "RGBA8",
-                "dimensionType": "ScreenRelative",
-                "width": 1,
-                "height": 1
-            }
-        },
-        {
-            "name": "DepthBuffer",
-            "format": {
-                "pixelFormat": "Depth",
-                "dimensionType": "ScreenRelative",
-                "width": 1,
-                "height": 1
-            }
-        }
-    ],
-    "samplers": [
-        {
-            "name": "Point",
-            "filter": "Point",
-            "wrapMode": "Clamp"
-        }
-    ]
-}
